arm/ipipe: fix off-by-one last IPI number
Although the regular SMP code is fine with keeping IPI_CPU_BACKTRACE
out of this count, we don't as we depend on it to figure out the exact
number of IPI vectors in use. The last IPI number we must allocate is
NR_IPI + IPIPE_LAST_IPI + 1, to account for the empty CPU_BACKTRACE
slot.

This bug badly breaks ipipe_raise_irq() by causing a conflict in the
VIRQ namespace between the last IPI and the first allocated VIRQ
allocated by the generic code (detected with deferred out-of-band
printk() messages not showing up).